backend/requirements.txt
# 财务付款对比系统 - 后端依赖
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.3
mysqlclient>=2.2
redis>=4.5
celery>=5.3
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
openpyxl>=3.1
reportlab>=4.0
requests>=2.31